+++ R/pprimarycensored.R
@@ -8,7 +8,10 @@
 #'
 #' @param q Vector of quantiles
 #'
-#' @param pdist Distribution function (CDF)
+#' @param pdist Distribution function (CDF). The package can identify base R
+#' distributions for potential analytical solutions. For non-base R functions,
+#' users can apply [add_name_attribute()] to yield properly tagged
+#' functions if they wish to leverage the analytical solutions.
 #'
 #' @param pwindow Primary event window
 #'
@@ -21,24 +24,23 @@
 #' normalized to integrate to 1 over \[0, pwindow\]. Defaults to a uniform
 #' distribution over \[0, pwindow\]. Users can provide custom functions or use
 #' helper functions like `dexpgrowth` for an exponential growth distribution.
-#' See `primary_dists.R` for examples.
+#' See `primary_dists.R` for examples. The package can identify base R
+#' distributions for potential analytical solutions. For non-base R functions,
+#' users can apply [add_name_attribute()] to yield properly tagged
+#' functions if they wish to leverage analytical solutions.
 #'
 #' @param dprimary_args List of additional arguments to be passed to
 #' dprimary. For example, when using `dexpgrowth`, you would
 #' pass `list(min = 0, max = pwindow, r = 0.2)` to set the minimum, maximum,
 #' and rate parameters
 #'
-#' @param pdist_name A string specifying the name of the delay distribution
-#' function. If NULL, the function name is extracted using
-#' [.extract_function_name()]. Used to determine if a analytical solution
-#' exists for the primary censored distribution. Must be set if `pdist` is
-#' passed a pre-assigned variable rather than a function name.
+#' @param pdist_name `r lifecycle::badge("deprecated")` this argument will be
+#' ignored in future versions; use [add_name_attribute()] on `pdist`
+#' instead
 #'
-#' @param dprimary_name A string specifying the name of the primary event
-#' distribution function. If NULL, the function name is extracted using
-#' [.extract_function_name()]. Used to determine if a analytical solution
-#' exists for the primary censored distribution. Must be set if `dprimary` is
-#' passed a pre-assigned variable rather than a function name.
+#' @param dprimary_name `r lifecycle::badge("deprecated")` this argument will be
+#' ignored in future versions; use [add_name_attribute()] on `dprimary`
+#' instead
 #'
 #' @param ... Additional arguments to be passed to pdist
 #'
@@ -99,24 +101,26 @@
 #' )
 pprimarycensored <- function(
     q, pdist, pwindow = 1, D = Inf, dprimary = stats::dunif,
-    dprimary_args = list(), pdist_name = NULL, dprimary_name = NULL, ...) {
+    dprimary_args = list(),
+    pdist_name = lifecycle::deprecated(),
+    dprimary_name = lifecycle::deprecated(),
+    ...) {
+  nms <- .name_deprecation(pdist_name, dprimary_name)
+  if (!is.null(nms$pdist)) {
+    pdist <- add_name_attribute(pdist, nms$pdist)
+  }
+  if (!is.null(nms$dprimary)) {
+    dprimary <- add_name_attribute(dprimary, nms$dprimary)
+  }
+
   check_pdist(pdist, D, ...)
   check_dprimary(dprimary, pwindow, dprimary_args)
-
-  if (is.null(pdist_name)) {
-    pdist_name <- .extract_function_name(substitute(pdist))
-  }
-  if (is.null(dprimary_name)) {
-    dprimary_name <- .extract_function_name(substitute(dprimary))
-  }
 
   # Create a new primarycensored object
   pcens_obj <- new_pcens(
     pdist,
     dprimary,
     dprimary_args,
-    pdist_name = pdist_name,
-    dprimary_name = dprimary_name,
     ...
   )