+++ NEWS.md
@@ -1,3 +1,4 @@
+
 # primarycensored 1.0.0.1000
 
 Development release.
@@ -5,6 +6,7 @@ Development release.
 ## Package
 
 - Updated the CI so that stan code is also tested on Windows and Mac. This is to ensure that the stan code is compatible with these platforms and in response to a CI bug in `epidist`.
+- Revised approach to automatic discovery of distribution functions. This soft deprecates the `pdist_name` and `dprimary_name` arguments throughout. Users wishing to pass distribution names (i.e. to potentially leverage analytical solutions) are advised to use the newly introduced `add_name_attribute()` function. Adds transient dependency on `lifecycle` and `rlang` packages. See #188 by @pearsonca.
 
 ## Documentation